feat(reimbursement): modified controller
